requirements.txt
numpy==1.11.1
ruamel.yaml==0.11.15
nltk==3.2.1
